requirements.txt
Django==3.2.5
djangorestframework==3.12.4
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==3.2.5
django-cors-headers==4.3.1
djangorestframework==3.12.4
idna==3.4
psycopg2-binary==2.9.9
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
